requirements.txt
# Voting / Meeting Q&A app
fastapi>=0.100
uvicorn[standard]>=0.20
sqlalchemy[asyncio]>=2.0
python-multipart>=0.0.6
jinja2>=3.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
httpx>=0.24
itsdangerous>=2.0
aiosqlite>=0.19
greenlet>=3.0
cryptography>=41
python-dotenv>=1.0
